requirements.txt
discord.py==2.3.4
transformers==4.30.2
torch
detoxify==0.4.1